Documentation fixes; minor rearrangement. Docs are a bit unorganized and need some additional work.
