+++ docs/endaq/device/command_interface.rst
@@ -1,0 +1,7 @@
+#################
+Command Interface
+#################
+
+.. automodule:: endaq.device.command_interfaces
+  :members:
+  :noindex:
