+++ docs/endaq/device/config.rst
@@ -1,0 +1,7 @@
+#############
+Configuration
+#############
+
+.. autoclass:: endaq.device.config
+  :members:
+  :noindex:
