+++ docs/endaq/device/mqtt.rst
@@ -1,0 +1,10 @@
+#####################
+``endaq.device.mqtt``
+#####################
+
+.. automodule:: endaq.device.mqtt
+  :members:
+  :imported-members:
+
+.. automodule:: endaq.device.mqtt.manager
+  :members:
